requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
PyMuPDF==1.23.5
python-docx==0.8.11
tqdm==4.66.1
pytesseract==0.3.10
Pillow==10.0.1
werkzeug==2.3.7
